+++ src/Tests/Functions.fs
@@ -14,6 +14,7 @@ let getResource<'T when 'T :> IArmResource> (data:IArmResource list) = data |> L
 /// Client instance needed to get the serializer settings.
 let dummyClient = new WebSiteManagementClient (Uri "http://management.azure.com", TokenCredentials "NotNullOrWhiteSpace")
 let getResourceAtIndex o = o |> getResourceAtIndex dummyClient.SerializationSettings
+let getResources (v:IBuilder) = v.BuildResources Location.WestEurope
 
 let tests = testList "Functions tests" [
     test "Renames storage account correctly" {
@@ -74,5 +75,35 @@ let tests = testList "Functions tests" [
 
         let f:Site = functions { worker_process Bitness.Bits64 } |> getResourceAtIndex 0
         Expect.equal f.SiteConfig.Use32BitWorkerProcess (Nullable false) "Should not use 32 bit worker process"
+    }
+
+    test "Managed KV integration works correctly" {
+        let sa = storageAccount { name "teststorage" }
+        let wa = functions { name "testfunc"; setting "storage" sa.Key; secret_setting "secret"; setting "literal" "value"; link_to_keyvault (ResourceName "testfuncvault") }
+        let vault = keyVault { name "testfuncvault"; add_access_policy (AccessPolicy.create (wa.SystemIdentity.PrincipalId, [ KeyVault.Secret.Get ])) }
+        let vault = vault |> getResources |> getResource<Vault> |> List.head
+        let secrets = wa |> getResources |> getResource<Vaults.Secret>
+        let site = wa |> getResources |> getResource<Web.Site> |> List.head
+
+        let expectedSettings = Map [
+            "storage", LiteralSetting "@Microsoft.KeyVault(SecretUri=https://testfuncvault.vault.azure.net/secrets/storage)"
+            "secret", LiteralSetting "@Microsoft.KeyVault(SecretUri=https://testfuncvault.vault.azure.net/secrets/secret)"
+            "literal", LiteralSetting "value"
+        ]
+
+        Expect.equal site.Identity.SystemAssigned Enabled "System Identity should be enabled"
+        Expect.containsAll site.AppSettings expectedSettings "Incorrect settings"
+
+        Expect.equal wa.Identity.SystemAssigned Enabled "System Identity should be turned on"
+
+        Expect.hasLength secrets 2 "Incorrect number of KV secrets"
+
+        Expect.equal secrets.[0].Name.Value "testfuncvault/secret" "Incorrect secret name"
+        Expect.equal secrets.[0].Value (ParameterSecret (SecureParameter "secret")) "Incorrect secret value"
+        Expect.sequenceEqual secrets.[0].Dependencies [ vaults.resourceId "testfuncvault" ] "Incorrect secret dependencies"
+
+        Expect.equal secrets.[1].Name.Value "testfuncvault/storage" "Incorrect secret name"
+        Expect.equal secrets.[1].Value (ExpressionSecret sa.Key) "Incorrect secret value"
+        Expect.sequenceEqual secrets.[1].Dependencies [ vaults.resourceId "testfuncvault"; storageAccounts.resourceId "teststorage" ] "Incorrect secret dependencies"
     }
 ]